data types fix

+++ src/main/java/com/onlol/fetcher/api/model/ApiChampionInfoDTO.java
@@ -7,23 +7,23 @@ import java.util.List;
 /lol/platform/v3/champion-rotations
  */
 public class ApiChampionInfoDTO {
-    private List<Integer> freeChampionIdsForNewPlayers = new ArrayList<>();
-    private List<Integer> freeChampionIds = new ArrayList<>();
+    private List<Long> freeChampionIdsForNewPlayers = new ArrayList<>();
+    private List<Long> freeChampionIds = new ArrayList<>();
     private Integer maxNewPlayerLevel = 0;
 
-    public List<Integer> getFreeChampionIdsForNewPlayers() {
+    public List<Long> getFreeChampionIdsForNewPlayers() {
         return freeChampionIdsForNewPlayers;
     }
 
-    public void setFreeChampionIdsForNewPlayers(List<Integer> freeChampionIdsForNewPlayers) {
+    public void setFreeChampionIdsForNewPlayers(List<Long> freeChampionIdsForNewPlayers) {
         this.freeChampionIdsForNewPlayers = freeChampionIdsForNewPlayers;
     }
 
-    public List<Integer> getFreeChampionIds() {
+    public List<Long> getFreeChampionIds() {
         return freeChampionIds;
     }
 
-    public void setFreeChampionIds(List<Integer> freeChampionIds) {
+    public void setFreeChampionIds(List<Long> freeChampionIds) {
         this.freeChampionIds = freeChampionIds;
     }
